Cool changes.
Firing sounds added.

Zombies spawn from 4 corners of the map too.

Zombies have now 10% chance of spawning with about triple movespeed.
They can be our fast movers. Just to demonstrate. Code to alter this is
in ZombieMover.

+++ Assets/Scripts/Zombie/ZombieHealthController.js
@@ -1,7 +1,7 @@
 ﻿#pragma strict
 
-var startHealth : int;
-private var health : int;
+var startHealth : float;
+private var health : float;
 private var animatorDead : boolean;
 
 var angleDeviationOfDying : float;
